requirements.txt
numpy==1.13.3
scipy==1.0.0
pandas==0.20.1
scikit-learn==0.19.0
matplotlib==2.1.2
tensorflow==1.3.0
Keras==2.0.8 
h5py==2.7.1
Django==1.11.6
djangorestframework==3.7.3
textblob-de==0.4.2
gensim==3.1.0
